Move isolated_cache() decorator to utils file
